pyproject.toml
[project]
name = "llm-api"
version = "0.1.0"
description = "Api connecting all the Free LLM APIs on the web"
authors = [
    {name = "Ujwal",email = "ujwalujwalc@gmail.com"}
]
readme = "README.md"
requires-python = "3.13.3"
dependencies = [
    "fastapi>=0.115.14,<0.116.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "httpx>=0.25.0",
    "aiohttp>=3.9.0",
    "python-dotenv>=1.0.0",

    "asyncio-throttle>=1.0.2",
    "tenacity>=8.2.0",
    "python-multipart>=0.0.6"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false